+++ src/main/java/ai/jgp/gha/dataproduct/K.java
@@ -5,7 +5,7 @@ package ai.jgp.gha.dataproduct;
  */
 public class K {
 
-    public static final String VERSION = "0.4.1";
+    public static final String VERSION = "0.4.2";
 
     public static final String DEFAULT_CATALOG = "default";